backend/requirements.txt
asgiref==3.11.0
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
et_xmlfile==2.0.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
